Rename folders; Moved files; Changed Menu Component
