+++ pages/components/GlobalStyle.js
@@ -1,0 +1,28 @@
+import {createGlobalStyle} from 'styled-components';
+
+const GlobalStyle = createGlobalStyle`
+  html,
+  body {
+    height: 100%;
+  }
+
+  body {
+    background-color: ${props => props.theme.primary};
+    font-family: 'Source Sans Pro', sans-serif;
+    transition-property: background-color;
+    transition-duration: 300ms;
+    margin: 0;
+    padding: 0;
+  }
+
+  #__next {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+  }
+`;
+
+export default GlobalStyle;
