+++ pages/components/ThemeComponents/ThemeColorschemes.js
@@ -1,11 +1,13 @@
 const darkTheme = {
   primary: '#07244C',
   secondary: '#ffffff',
+  orange: '#FAA21B'
 };
 
 const lightTheme = {
   primary: '#ffffff',
   secondary: '#07244C',
+  orange: '#FAA21B',
 };
 
 const getTheme = currentTheme => {
